backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-decouple
psycopg2-binary
python-dotenv
django-phonenumber-field[phonenumbers]
google-auth
django-cors-headers

